+++ modules/core-config/modules/loki/variables.tf
@@ -19,6 +19,12 @@ variable "cluster_name" {
   nullable    = false
 }
 
+variable "cluster_version" {
+  description = "The Kubernetes version to use for the EKS cluster."
+  type        = string
+  nullable    = false
+}
+
 variable "cluster_oidc_issuer_url" {
   description = "The OIDC issuer url for the cluster."
   type        = string