+++ modules/core-config/modules/node-config/variables.tf
@@ -1,3 +1,9 @@
+variable "cluster_version" {
+  description = "The Kubernetes version to use for the EKS cluster."
+  type        = string
+  nullable    = false
+}
+
 variable "namespace" {
   description = "Namespace to install the Kubernetes resources into."
   type        = string